move field definitions to common_fields.py
